Added New Features as follows
Added a timed automatic script shutdown to kill terminal in preparation for anautomatic system shutdown. Variable time intervals to check what hour it is and when it is equal to the hour for shutdown. 15min 30min 1hr. The shutdown hour is in the 24hr format 1-24 are selectable.

Had a bit of a hackish way of changing the delay that dealt with invalidating the timer and then remaking it with the new delay. This of course could be resource intensive but as there is not a direct way to change the interval it will have to do.